✏️ コンパイラ警告の対応

+++ src/md.cpp
@@ -531,7 +531,7 @@ void MD::run(void) {
     if (mi.rank == 0) {
         for (const auto & file : std::filesystem::directory_iterator("./")) {
             std::string path = file.path();
-            int word_pos = path.find("/cdv");
+            size_t word_pos = path.find("/cdv");
             if (word_pos != std::string::npos) {
                 std::filesystem::remove_all(path);
                 continue;
